removed batcher, replaced with simple loop algorithm
