+++ src/Literature/DiscoverableScriptContents/hack.wip.js
@@ -1,6 +1,8 @@
 /** @param {NS} ns */
 export async function main(ns) {
   ns.disableLog('ALL');
-  const [target, additionalMsec] = ns.args;
-  await ns.hack(target, { additionalMsec });
+  const [target] = `${ns.args}`;
+  while(true) {
+    await ns.hack(target);
+  }
 }
